% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpectra.R
\name{ArraySpectra}
\alias{ArraySpectra}
\title{Spectral estimates of core array}
\usage{
ArraySpectra(cores, res = 1, neff = length(cores), df.log = NULL,
  pval = 0.05, ...)
}
\arguments{
\item{cores}{a list of the proxy data from the core array. Each component is
expected to be a numeric vector of the proxy values of a common length.}

\item{res}{the sampling (e.g., temporal) resolution of the proxy data;
determines the frequency axis of the spectral estimates.}

\item{neff}{the effective number of records (e.g. to account for an expected
spatial correlation of the local noise). Per default, no spatial correlation
is assumed and \code{neff} is set to the number of proxy records (the length
of \code{cores}).}

\item{df.log}{width of the Gaussian kernel in logarithmic
space to smooth the spectral estimates; \code{NULL} (the default) suppresses
smoothing. Log-smoothing automatically results in calculation of
confidence intervals using a p-value of \code{0.05}.}

\item{pval}{p-value for calculating confidence intervals of the spectral
estimates if no log-smoothing is applied to the spectra, or for
re-calculating the confidence intervals with a different p-value than the
default \code{0.05}.}

\item{...}{additional parameters supplied to the spectral estimation
function.}
}
\value{
A list of the following components:
\describe{
\item{N:}{the number of (effective) proxy records of the core array.}
\item{single:}{a list of length \code{N} with the spectra of each individual
proxy record.}
\item{mean:}{the mean spectrum across all individual spectra.}
\item{stack:}{the spectrum of the average proxy record in the time domain
("stacked record").}
}
}
\description{
\code{ArraySpectra} calculates all relevant spectral estimates
from a given core array of \code{N} proxy records. All spectral estimates
can be smoothed in logarithmic space and confidence intervals can be
calculated.
}
\details{
The spectral estimates are calculated using the function
\code{SpecMTM} from the package \code{PaleoSpec}, which applies the
multi-taper method. If the package is not available, spectra are
calculated using base \code{R} functions, which may result in estimates with
higher spectral uncertainty. In both cases, each spectral result is returned
as a \code{"spectral object"} list with the minimum elements \code{freq} and
\code{spec}. Log-smoothing and (re-)calculation of confidence intervals is
only available with \code{PaleoSpec}.

\code{PaleoSpec} is available upon request.
}
\seealso{
\code{[PaleoSpec]{SpecMTM}}
}
\author{
Thomas Münch
}
