% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clim.site.par}
\alias{clim.site.par}
\title{Climatological site parameters}
\format{
A data frame with 20 rows and 5 variables:
\describe{
  \item{core:}{the name of the firn-core site}
  \item{temperature:}{10-m firn temperature (or annual mean air
    temperature), in degree Celsius}
  \item{acc.rate:}{accumulation rate, in kg/m^2/yr}
  \item{pressure:}{surface air pressure, in mbar}
  \item{height:}{surface elevation, in m}
}
}
\source{
All values for firn cores in the rows 1-15 from Oerter et al. (2000) (Table
1, 4), except: temperature for FB9807 adapted from B32 (1 km), temperature
for FB9815 adapted from FB9805 (28 km), temperature for FB9811 adapted from
FB9812 (19 km).

For firn cores in the rows 16-20, accumulation rates and elevation data are
for core WDC2005A taken from WAIS Divide Project Members (2013) (p.440) and
for the other cores from Kaspari et al. (2004) (Table 1), respectively.
WDC2005A temperature is taken from WAIS Divide Project Members (2013)
(p.440), the other temperature values are based on ERA-Interim mean
temperature anomalies (Dee et al., 2011) with respect to WDC2005A.

For all cores, the pressure values are calculated with the barometric formula
using the local elevation and temperature data.
}
\usage{
clim.site.par
}
\description{
A dataset containing the climatological site parameters that are relevant for
the diffusion length calculation for all firn-core sites used in Münch and
Laepple (2018).
}
\references{
Dee, D. P., et al.: The ERA-Interim reanalysis: configuration and performance
of the data assimilation system, Q. J. R. Meteorol. Soc., 137(656), 553–597,
\url{https://doi.org/10.1002/qj.828}, 2011.

Kaspari, S., et al.: Climate variability in West Antarctica derived from
annual accumulation-rate records from ITASE firn/ice cores, Ann. Glaciol.,
39(1), 585-594, \url{https://doi.org/10.3189/172756404781814447}, 2004.

Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, \url{https://doi.org/10.5194/cp-14-2053-2018},
2018.

Oerter, H., et al.: Accumulation rates in Dronning Maud Land, Antarctica, as
revealed by dielectric-profiling measurements of shallow firn cores,
Ann. Glaciol., 30(1), 27-34, 2000.

WAIS Divide Project Members: Onset of deglacial warming in West Antarctica
driven by local orbital forcing, Nature, 500(7463), 440–444,
\url{https://doi.org/10.1038/nature12376}, 2013.
}
\keyword{datasets}
