% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-data.R
\docType{data}
\name{diffusion.length}
\alias{diffusion.length}
\title{Diffusion lengths}
\format{A list with 3 elements:
\describe{
  \item{dml1:}{a data frame with 338 rows and 16 variables:}
    \describe{
      \item{Time:}{age of the firn core, in yr}
      \item{columns "B31"-"FB9817":}{15 columns with 338 rows with diffusion
      lengths corresponding to the firn-core ages in \code{Time}, in yr}
  }
  \item{dml2:}{a data frame with 1990 rows and 4 variables:}
    \describe{
      \item{Time:}{age of the firn core, in yr}
      \item{columns "B31"-"B33":}{3 columns with 1990 rows with diffusion
      lengths corresponding to the firn-core ages in \code{Time}, in yr}
  }
  \item{wais:}{a data frame with 402 rows and 6 variables:}
    \describe{
      \item{Time:}{age of the firn core, in yr}
      \item{columns "WDC2005A"-"ITASE-2000-5":}{5 columns with 402 rows with
      diffusion lengths corresponding to the firn-core ages in \code{Time},
      in yr}
  }
}}
\usage{
diffusion.length
}
\description{
A dataset containing calculated diffusion lengths for all firn-core sites
used in Münch and Laepple (2018). The diffusion lengths are provided as a
function of firn-core age at semi-annual resolution and in units of years.
}
\details{
Diffusion lengths as a function of firn-core depth are calculated from
Eq. (8) in Gkinis et al. (2014); firn diffusivity is modelled following
Eqs. (17-19) in Johnsen et al. (2000), firn density with the Herron–Langway
model (Herron and Langway, 1980; Eqs. 7 and 10). For the surface firn density
we assume a constant value of 340 kg/m^3 for all sites, for all other
relevant site parameters, see \code{?clim.site.par}. The depth-dependent
diffusion length values are converted into units of yr using the
Herron-Langway firn density with a constant accumulation rate, related to
the firn-core age and then linearly interpolated onto an equidistant time
axis at semi-annual resolution. All steps are documented in the package
source (\code{data-raw/calculate-diffusion-length.R}).
}
\references{
Gkinis, V., et al.: Water isotope diffusion rates from the North-GRIP ice
core for the last 16,000 years – Glaciological and paleoclimatic
implications, Earth Planet. Sci. Lett., 405, 132–141,
\url{https://doi.org/10.1016/j.epsl.2014.08.022}, 2014.

Herron, M. M. and Langway, Jr., C. C.: Firn Densification: An Empirical
Model, J. Glaciol., 25(95), 373–385,
\url{https://doi.org/10.3189/S0022143000015239}, 1980.

Johnsen, S. J., et al.: Diffusion of stable isotopes in polar firn and ice:
the isotope effect in firn diffusion, in: Physics of Ice Core Records, edited
by Hondoh, T., vol. 159, pp. 121–140, Hokkaido University Press, Sapporo,
Japan, 2000.

Münch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., \url{https://doi.org/10.5194/cp-2018-112}, in review,
2018.
}
\keyword{datasets}
