% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffusion.tf}
\alias{diffusion.tf}
\title{Diffusion transfer functions}
\format{
A named list with the 3 elements \code{dml1}, \code{dml2}, and
  \code{wais}, where each element is a spectral object:
  \describe{
  \item{\code{freq}:}{the frequency axis in units of \code{1 / yr},}
  \item{\code{spec}:}{the transfer function value at each frequency.}
}
}
\source{
The transfer functions were obtained using \code{?CalculateDiffusionTF} for
the site-specific diffusion lengths provided by
\code{proxysnr:::diffusion.length}; see also the respective package
\code{vignette("calculate-transfer-functions")}. Isotope records were
simulated at semi-annual resolution and the transfer functions interpolated
in frequency space to annual resolution. Note here that prior to
\code{proxysnr} v1.0.0, the transer functions supplied to
\code{\link{SeparateSignalFromNoise}} needed to match in frequency axis the
spectra of the analysed data, hence the interpolation step. From
\code{proxysnr} v1.0.0 onwards this is no longer required, as the frequency
axis overlap is checked within \code{\link{SeparateSignalFromNoise}}, with
interpolation taking place there, if needed. The example firn diffusion
transfer functions here were, however, left as originally produced to
maintain consistency with the results published in Münch and Laepple (2018).
}
\usage{
diffusion.tf
}
\description{
A dataset containing the calculated diffusion transfer functions for the
DML1, DML2 and WAIS records from Münch and Laepple (2018) based on numerical
simulations. The transfer function describes the average effect firn
diffusion has on the mean of the spectra from a number of \code{N} spatially
distributed oxygen isotope records.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, \url{https://doi.org/10.5194/cp-14-2053-2018},
2018.
}
\seealso{
\code{\link{CalculateDiffusionTF}},
  \code{vignette("calculate-transfer-functions")}
}
\keyword{datasets}
