% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dml}
\alias{dml}
\title{DML oxygen isotope data}
\format{
A named list with two components:
\describe{
  \item{dml1:}{a named list with 15 components where each component is a
    vector of length 194 with the oxygen isotope data in per mil for all 15
    cores for the years from 1994 CE to 1801 CE.}
  \item{dml2:}{a named list with 3 components where each component is a vector
    of length 995 with the oxygen isotope data in per mil for the cores
    \code{"B31"}, \code{"B32"} and \code{"B33"} for the years from 1994 CE to
    1000 CE.}
}
}
\source{
Data citation:\cr
Graf, W., et al.: Stable-isotope records from Dronning Maud Land, Antarctica,
\url{https://doi.org/10.1594/PANGAEA.728240}, 2002.

Literature citation:\cr
Graf, W., et al.: Stable-isotope records from Dronning Maud Land, Antarctica,
Ann. Glaciol., 35, 195–201, 2002.\cr
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, \url{https://doi.org/10.5194/cp-14-2053-2018},
2018.
}
\usage{
dml
}
\description{
A dataset containing the oxygen isotope data of 15 firn cores from Dronning
Maud Land (DML), Antarctica, as used in Münch and Laepple (2018).
}
\details{
Data has been processed from the original source data as documented in the
package source (\code{data-raw/load.DML.R}). Cores included are FB9804,
FB9805, FB9807–FB9811, FB9813–FB9817, and B31–B33.
}
\keyword{datasets}
