% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCI.R
\name{EstimateCI}
\alias{EstimateCI}
\title{Confidence interval estimation}
\usage{
EstimateCI(
  spectra,
  f.start = NULL,
  f.end = NULL,
  nmc = 10,
  probs = c(0.1, 0.9),
  df.log = NULL,
  ci.df.log = NULL
)
}
\arguments{
\item{spectra}{a list with the spectral objects \code{signal}, \code{noise},
and \code{snr} for an investigated proxy record array, obtained from
\code{\link{SeparateSignalFromNoise}}.}

\item{f.start}{lower end of the frequency range on which the power-law fit is
made on the proxy data (see Details); the default \code{NULL} uses the
lowest frequency of the proxy \code{spectra}.}

\item{f.end}{as \code{f.start} for the upper end; the default \code{NULL}
uses the uppermost frequency of the proxy \code{spectra}.}

\item{nmc}{integer; the number of replications for the confidence interval
estimation.}

\item{probs}{length-2 numeric vector of probabilities with values in [0,1]
defining the confidence interval; defaults to the 10-90 \% interval.}

\item{df.log}{width of the Gaussian kernel in logarithmic frequency units to
smooth the spectral estimates of the simulated data; some smoothing is
usually necessary to avoid physically implausible negative power
occasionally occuring for some frequencies upon estimating the common
signal spectrum. It is suggested to use the same amount of smoothing as for
the actual proxy data, while setting \code{NULL} (the default) suppresses
smoothing.}

\item{ci.df.log}{width of the Gaussian smoothing kernel to smooth the
estimated confidence intervals, merely for visual purposes; \code{NULL}
(the default) suppresses smoothing.}
}
\value{
the input \code{spectra} object, amended by the confidence intervals
  for the signal, noise and SNR spectra (element \code{lim.1} gives the upper
  confidence level, element \code{lim.2} the lower level, respectively).
}
\description{
This function estimates based on a parametric bootstrapping procedure (see
Details) confidence intervals for a set of signal, noise and SNR spectra
obtained from an actual proxy record array.
}
\details{
The parametric bootstrapping procedure for the confidence level estimation is
implemented as follows. A power-law fit of the form \code{alpha * f^(-beta)}
is applied to the actual signal and noise spectra, and the resulting
power-law coefficients are used to generate surrogate signal and noise series
in a simulated array that mimics the actual proxy record array. This
simulated array is replicated \code{nmc} times and lower and upper quantiles
are calculated across the realizations of the signal, noise and SNR
surrogates. Subsequently, the quantiles are scaled to the respective mean
estimates and applied multiplicatively to the actual proxy estimates of
signal, noise, and SNR to yield the confidence intervals.
}
\seealso{
\code{\link{SeparateSignalFromNoise}},
  \code{\link{ObtainArraySpectra}}
}
\author{
Thomas Münch
}
