% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCI.R
\name{extractQuantiles}
\alias{extractQuantiles}
\title{Quantiles of simulated signal and noise spectrum realizations}
\usage{
extractQuantiles(surrogates, probs = c(0.1, 0.9), minval = 1e-15)
}
\arguments{
\item{surrogates}{the output from a run of \code{\link{runSurrogates}}.}

\item{probs}{length-2 numeric vector of probabilities with values in [0,1];
default setting extracts the 10 and 90 \% quantiles.}

\item{minval}{minimum value assigned to negative estimates of the scaled
lower and upper quantiles to avoid unphysical negative confidence
intervals, thereby preventing issues on logarithmic plots.}
}
\value{
a named list of the three elements `signal`, `noise`, and `snr`, with
  each list element in turn being a named list with the `lower` and `upper`
  quantiles (according to the setting of \code{probs}) across the simulated
  signal, noise and signal-to-noise ratio (SNR) spectra.
}
\description{
Extract lower and upper quantiles across the signal, noise and SNR
realizations from a run of \code{\link{runSurrogates}}, scaled to the
respective mean estimates of the realiztaions.
}
\seealso{
\code{\link{runSurrogates}}
}
\author{
Thomas Münch
}
\keyword{internal}
