% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIntegratedSNR.R
\name{GetIntegratedSNR}
\alias{GetIntegratedSNR}
\title{Frequency-dependent signal-to-noise ratio from integrated spectra}
\usage{
GetIntegratedSNR(input, N = 1, f1 = 2, f2 = "max", limits = NULL)
}
\arguments{
\item{input}{a list of the spectral objects lists \code{signal} and
\code{noise}, usually to be obtained from a call to
\code{\link{SeparateSignalFromNoise}}.}

\item{N}{integer; number of proxy records averaged. The default returns the
SNR of a single proxy record. For a different number, the SNR is calculated
for a "stack" averaged across \code{N} individual proxy records with the same
signal and equivalent noise characteristics, assuming independent noise
between the records.}

\item{f1}{index of the signal (and noise) frequency axis to specify the lower
integration limit from which to integrate the spectra; per default the
lowest frequency of the spectral estimates is omitted.}

\item{f2}{as \code{f1} for the upper integration limit; defaults to use the
maximum frequency of the given spectral estimates.}

\item{limits}{numeric vector with a frequency range of the integration: this
is an alternative way of specifying the integration limits and overrides
the setting by \code{f1} and \code{f2}. If not `NULL` it must be a length-2
vector with the lower integration limit as first and the upper integration
limit as second element.}
}
\value{
a spectral object list of the SNR.
}
\description{
This function calculates the signal-to-noise ratio (SNR) as a function of
frequency, interpreted as the temporal resolution of a proxy record. This
variant of the SNR is obtained from a signal and a noise spectrum which are
each integrated across frequencies before taking their ratio.
}
\details{
The function is an implementation of Eq. (6) in Münch and Laepple
(2018). The integral in (6) is approximated by the cumulative sum of the
integration arguments from \code{f.int1} to \code{f.int2}, where
\code{f.int1 = f1} and \code{f.int2} consecutively increases from \code{f1}
to \code{f2}.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\author{
Thomas Münch
}
