% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegratedSNR.R
\name{IntegratedSNR}
\alias{IntegratedSNR}
\title{Frequency-dependent signal-to-noise ratio from integrated spectra}
\usage{
IntegratedSNR(
  input,
  N = 1,
  f1 = 2,
  f2 = "max",
  freq.cut.lower = NULL,
  freq.cut.upper = NULL
)
}
\arguments{
\item{input}{a list of the spectral objects lists \code{signal} and
\code{noise}, usually to be obtained from a call to
\code{\link{SeparateSpectra}}}

\item{N}{integer; number of proxy records averaged. The default returns the
SNR assuming a single proxy record. For a different number, the SNR is
scaled by this number, assuming independent noise between the records.}

\item{f1}{index of the the minimum frequency from which to integrate the
signal and noise spectra for calculating the SNR; per default the
lowest frequency of the spectral estimates is omitted}

\item{f2}{index of the maximum frequency until which to integrate the signal
and noise spectra for calculating the SNR; defaults to use the
maximum frequency of the given spectral estimates}

\item{freq.cut.lower}{lower frequency (not index!) at which to cut the
spectra: this provides a direct way for specifying a minimum frequency for
the integration different from the minimum frequency of the spectral
estimates. Setting \code{freq.cut.lower} overrides the frequency
corresponding to the index set in \code{f1}.}

\item{freq.cut.upper}{upper frequency (not index!) at which to cut the
spectra: this provides a direct way for specifying a maximum frequency for
the integration different from the maximum frequency of the spectral
estimates. Setting \code{freq.cut.upper} overrides the frequency
corresponding to the index set in \code{f2}.}
}
\value{
a spectral object list of the SNR.
}
\description{
This function calculates the signal-to-noise ratio (SNR) as a function of
frequency, interpreted as the temporal resolution of a proxy record. This
variant of the SNR is obtained from the ratio of a signal and a noise
spectrum, which are integrated before cumulatively across frequencies.
}
\details{
The function is an implementation of Eq. (6) in Münch and Laepple
(2018). The integral in (6) is approximated by the cumulative sum of the
integration arguments from \code{f.int1} to \code{f.int2}, where
\code{f.int1 = f1} and \code{f.int2} consecutively increases from \code{f1}
to \code{f2}.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\author{
Thomas Münch
}
