% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-utils.R
\name{LLines}
\alias{LLines}
\title{Add spectral object plot}
\usage{
LLines(
  x,
  inverse = FALSE,
  conf = TRUE,
  col = "black",
  alpha = 0.2,
  removeFirst = 0,
  removeLast = 0,
  ...
)
}
\arguments{
\item{x}{a spectral object.}

\item{inverse}{if \code{TRUE} treat the x-axis values in units of period
(inverse frequency), increasing to the left; defaults to \code{FALSE}.}

\item{conf}{if \code{TRUE} (the default) add a transparent confidence
interval (has no effect if \code{x} contains no confidence limits).}

\item{col}{color for the line plot and the confidence interval.}

\item{alpha}{transparency level (between 0 and 1) for the confidence
interval; defaults to \code{0.2}.}

\item{removeFirst}{omit \code{removeFirst} values on the low frequency side.}

\item{removeLast}{omit \code{removeLast} values on the high frequency side.}

\item{...}{further graphical parameters passed to \code{lines}.}
}
\description{
Add a line plot of a spectral object to an existing plot.
}
\seealso{
`?spec.object` for the definition of a "proxysnr" spectral object.
}
\author{
Thomas Laepple, Thomas Münch
}
