% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-estimation.R
\name{LogSmooth}
\alias{LogSmooth}
\title{Spectral smoothing}
\usage{
LogSmooth(x, df.log = 0.05)
}
\arguments{
\item{x}{the spectrum as an object of class \code{"spec"} with the degrees
of freedom of the spectrum as an additional list element \code{dof}; or
simply a list with elements \code{spec}, \code{freq} and \code{dof}.}

\item{df.log}{smoothing width in logarithmic frequency units.}
}
\value{
an object of class \code{"spec"} with the smoothed spectrum
including the average degrees of freedom.
}
\description{
Smooth spectrum using Gaussian kernel with constant width in logarithmic
frequency space.
}
\author{
Thomas Laepple
}
