% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{LPlot}
\alias{LPlot}
\title{Log-log spectral plot}
\usage{
LPlot(
  x,
  conf = TRUE,
  bPeriod = FALSE,
  bNoPlot = FALSE,
  axes = TRUE,
  col = "black",
  alpha = 0.2,
  removeFirst = 0,
  removeLast = 0,
  xlab = "f",
  ylab = "PSD",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"spec"}.}

\item{conf}{if \code{TRUE} (the default) add a transparent confidence
interval (suppressed if \code{x} contains no error limits).}

\item{bPeriod}{if \code{TRUE} the x-axis is displayed in units of period
(inverse frequency), increasing to the left. Defaults to \code{FALSE}.}

\item{bNoPlot}{if \code{TRUE} only produce the plot frame (\code{type = "n"}
behaviour of function \code{\link{plot}}). Defaults to \code{FALSE}.}

\item{axes}{if \code{FALSE} the plotting of the x and y axes is
suppressed. Defaults to \code{TRUE}.}

\item{col}{color for the line plot and the confidence interval.}

\item{alpha}{transparency level (between 0 and 1) for the confidence
interval. Defaults to \code{0.2}.}

\item{removeFirst}{omit \code{removeFirst} values on the low frequency side.}

\item{removeLast}{omit \code{removeLast} values on the high frequency side.}

\item{xlab}{character string for labelling the x-axis.}

\item{ylab}{character string for labelling the y-axis.}

\item{xlim}{range of x-axis values; if \code{NULL} (the default) it is
calculated internally and automatically reversed for \code{bPeriod = TRUE}.}

\item{ylim}{range of y-axis values; if \code{NULL} (the default) it is
calculated internally.}

\item{...}{further graphical parameters passed to \code{plot}.}
}
\description{
This function plots a spectrum on a double-logarithmic scale and optionally
adds a transparent confidence interval.
}
\author{
Thomas Laepple
}
