% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-estimation.R
\name{MeanSpectrum}
\alias{MeanSpectrum}
\title{Mean spectrum}
\usage{
MeanSpectrum(speclist)
}
\arguments{
\item{speclist}{a list of objects of class \code{"spec"}, or lists with
minimum components \code{freq}, \code{spec} and \code{dof}, that supply the
spectra which are to be averaged.}
}
\value{
an object of class \code{"spec"} with components \code{freq},
\code{spec} and \code{dof}.
}
\description{
Calculate the mean spectrum of all supplied individual spectra.
}
\details{
The mean spectrum is calculated as a simple average across all power
spectral densities from the individual spectra. Degrees of freedom
(\code{dof}) of the mean spectrum are obtained from the sum of the
individual dof's at each frequency.
}
\author{
Thomas Laepple, Thomas Münch
}
