% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{muench_laepple_fig02}
\alias{muench_laepple_fig02}
\title{Figure 2 in Münch and Laepple (2018)}
\usage{
muench_laepple_fig02(spec, f.cut = FALSE)
}
\arguments{
\item{spec}{output from \code{\link{WrapSpectralResults}}.}

\item{f.cut}{Shall the spectra be cut at the cutoff frequency constrained
by the diffusion correction strength? Defaults to \code{FALSE}.}
}
\description{
Plot the signal and noise spectra from the DML and WAIS oxygen isotope data
sets (Figure 2 in Münch and Laepple, 2018).
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{WrapSpectralResults}}
}
\author{
Thomas Münch
}
