% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{muench_laepple_fig05}
\alias{muench_laepple_fig05}
\title{Figure 5 in M端nch and Laepple (2018)}
\usage{
muench_laepple_fig05(SNR, TNS, f.cut = FALSE)
}
\arguments{
\item{SNR}{output from \code{\link{PublicationSNR}}.}

\item{TNS}{output from \code{\link{TrenchNoise}}.}

\item{f.cut}{Shall the spectra be cut at the cutoff frequency constrained
by the diffusion correction strength? Defaults to \code{FALSE}.}
}
\description{
Plot comparison of the DML and trench noise spectra (Figure 5 in M端nch and
Laepple, 2018).
}
\references{
M端nch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
}
\seealso{
\code{\link{PublicationSNR}}, \code{\link{TrenchNoise}}
}
\author{
Thomas M端nch
}
