% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotStackCorrelation}
\alias{PlotStackCorrelation}
\title{Plot proxy stack correlation}
\usage{
PlotStackCorrelation(
  data,
  col.pal = NULL,
  label = "",
  xlim = NA,
  ylim = NA,
  xlab = "Number of cores",
  ylab = "Averaging period (yr)",
  xtm = NULL,
  ytm = NULL,
  xtl = NULL,
  ytl = NULL,
  xtm.min = NULL,
  ytm.min = NULL
)
}
\arguments{
\item{data}{a list of the correlation data (e.g. as output by
\code{\link{ObtainStackCorrelation}}), which must have two elements:
\code{freq} and \code{correlation}, where \code{freq} contains the
frequency axis of the underlying proxy data set (to obtain an axis for the
temporal averaging period), and where \code{correlation} is a \code{n * m}
matrix of the correlation values. The number of columns of
\code{correlation} must match the number of frequency values (i.e. the
length of \code{freq}), and the row index stands for the number of proxy
records averaged.}

\item{col.pal}{a color palette function to be used to assign colors in the
plot; the default `NULL` means to calculate the palette function internally
from ten colours of the diverging \code{RdYlBu} palette in the ColorBrewer
2.0 collection.}

\item{label}{an optional label of the data set to be displayed at the top
of the plot.}

\item{xlim}{the x limits (x1, x2) of the plot. Set to \code{NA} to use
default limits calculated from the x data range, or supply a numeric vector
of length 2 with custom limits. In the latter case, setting either of the
elements to \code{NA} results in using the default limit for this element
only; see the example.}

\item{ylim}{as \code{xlim} for the y limits of the plot.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{xtm}{x axis tick mark positions; default setting (\code{NULL}) uses
\code{c(1, 2, 5, 10, 20)}.}

\item{ytm}{y axis tick mark positions; default setting (\code{NULL}) uses
\code{c(2, 5, 10, 20, 50)}.}

\item{xtl}{x axis tick mark labels; if \code{NULL} determined automatically
from \code{xtm}, else it must be a vector of labels of the same length as
\code{xtm}.}

\item{ytl}{equivalent to \code{xtl} for the y axis tick mark labels.}

\item{xtm.min}{x axis minor tick marks; default setting \code{NULL} uses
minor tick marks that are adapted to the default x major tick marks. Set to
\code{NA} to omit minor ticks at all.}

\item{ytm.min}{as \code{xtm.min} for minor y axis tick marks.}
}
\description{
Plot the correlation of the spatial average of a certain number of proxy
records with the underlying common signal depending on the number of records
averaged and their temporal resolution (e.g., as in the firn core analysis of
Münch and Laepple, 2018, Fig. 4).
}
\examples{

# create a toy correlation dataset, which mimicks an increase
# in correlation with timescale and with the number of cores
# averaged, and plot it:

nf <- 20
nc <- 5

data <- list(
  freq = seq(0.01, 0.5, length.out = nf),
  correlation =
    matrix(seq(0.05, 0.9, length.out = nf), nrow = nc, ncol = nf,
           byrow = TRUE) +
    matrix(seq(0, 0.3, length.out = nc), nrow = nc, ncol = nf) +
    matrix(rnorm(nf * nc, sd = 0.02), nrow = nc, ncol = nf)
)
data$correlation[data$correlation > 1] <- 1
data$correlation[data$correlation < 0] <- 0

PlotStackCorrelation(data)


}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{ObtainStackCorrelation}}; <https://colorbrewer2.org/>
}
\author{
Thomas Münch
}
