% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{proxysnr-deprecated}
\alias{proxysnr-deprecated}
\title{Deprecated functions}
\description{
The following function names are deprecated, since they do not adhere to the
paradigm of using verb phrases, or are not desriptive enough, and are
replaced by the respective new names mentioned thereafter:

\itemize{
  \item \code{ArraySpectra} -> \code{\link{ObtainArraySpectra}}
  \item \code{SeparateSpectra} -> \code{\link{SeparateSignalFromNoise}}
  \item \code{StackCorrelation} -> \code{\link{ObtainStackCorrelation}}
  \item \code{IntegratedSNR} -> \code{\link{GetIntegratedSNR}}
  \item \code{DiffusionTF} -> \code{\link{CalculateDiffusionTF}}
  \item \code{TimeUncertaintyTF} -> \code{\link{CalculateTimeUncertaintyTF}}
}

However, calls that use the old name will still work and pipe the input to
the respective function of the new name, but a warning will be issued
informing about the new function name.
}
