% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxysnr-package.R
\docType{package}
\name{proxysnr-package}
\alias{proxysnr}
\alias{proxysnr-package}
\title{proxysnr: Separate Signal and Noise in Climate Proxy Records}
\description{
This R package implements a method to separate the common signal from local noise in a spatial network of climate proxy records using spectral analyses. The package allows the correction of the estimated signal and noise spectra for the effects of time uncertainty and diffusion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EarthSystemDiagnostics/proxysnr}
  \item Report bugs at \url{https://github.com/EarthSystemDiagnostics/proxysnr/issues}
}

}
\author{
\strong{Maintainer}: Thomas Münch \email{thomas.muench@awi.de}

Authors:
\itemize{
  \item Thomas Laepple
}

Other contributors:
\itemize{
  \item Andrew Dolman [contributor]
}

}
\keyword{internal}
