% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting.R
\name{PublicationSNR}
\alias{PublicationSNR}
\title{Final DML and WAIS spectra}
\usage{
PublicationSNR(
  spec,
  data = c("corr.full", "corr.diff.only", "corr.t.unc.only", "raw"),
  dml.knit.f = 0.1,
  df.log = 0.125
)
}
\arguments{
\item{spec}{spectral results for the DML and WAIS firn core data as output
from \code{\link{WrapSpectralResults}}.}

\item{data}{character string naming the version of the spectral results in
\code{spec} to use; one of "corr.full", "corr.diff.only",
"corr.t.unc.only", and "raw"; defaults to "corr.full".}

\item{dml.knit.f}{frequency at which to combine the spectra from the DML1
and DML2 data sets (defaults to 1/10 yr^(-1)); DML2 spectra are used for
lower frequencies than \code{dml.knit.f}, DML1 for the higher frequencies.}

\item{df.log}{width of Gaussian kernel in logarithmic frequency units for
additional smoothing for visual purposes; \code{NULL} suppresses smoothing.}
}
\value{
A list with the components \code{dml} and \code{wais}, where each is
  a list containing three spectral objects (`?spec.object`) and one numeric
  vector:
  \describe{
  \item{\code{signal}:}{the signal spectrum.}
  \item{\code{noise}:}{the noise spectrum.}
  \item{\code{snr}:}{the frequency-dependent signal-to-noise ratio.}
  \item{\code{f.cutoff}:}{a two-element vector with the index and value of
    the cutoff frequency from constraining the diffusion correction.}
}
}
\description{
Internal function to produce the final DML and WAIS spectra (signal, noise,
SNR) used for Figs. (3) and (4) in Münch and Laepple (2018). The DML spectra
are a combination of the results from the DML1 and DML2 data sets.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, \url{https://doi.org/10.5194/cp-14-2053-2018},
  2018.
}
\author{
Thomas Münch
}
\keyword{internal}
