% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCI.R
\name{runSurrogates}
\alias{runSurrogates}
\title{Monte Carlo signal and noise simulations}
\usage{
runSurrogates(signal.par, noise.par, nc, nt, res, nmc = 10, df.log = NULL)
}
\arguments{
\item{signal.par}{a named list with elements `alpha` and `beta` supplying the
power-law parameters to simulate the common signal.}

\item{noise.par}{as \code{signal.par} the power-law parameters to simulate
the independent noise components.}

\item{nc}{the number of proxy records (`"cores"`) in the array.}

\item{nt}{the length of each record in the array.}

\item{res}{the sampling (e.g., temporal) resolution of the simulated proxy
records; determines the frequency axis of the spectral estimates.}

\item{nmc}{integer; the number of replications.}

\item{df.log}{width of the Gaussian kernel in logarithmic frequency units to
smooth the spectral estimates; \code{NULL} (the default) suppresses
smoothing.}
}
\value{
a list of length \code{nmc} where each list element is one signal and
  noise estimation realization, i.e. the output of
  \code{\link{simSignalAndNoise}}.
}
\description{
Replicate \code{\link{simSignalAndNoise}} a given number of times.
}
\seealso{
\code{\link{simSignalAndNoise}}, \code{\link{simCoreArray}}
}
\author{
Thomas Münch
}
\keyword{internal}
