% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCI.R
\name{simCoreArray}
\alias{simCoreArray}
\title{Simulate a proxy core array}
\usage{
simCoreArray(signal.par, noise.par, nc, nt)
}
\arguments{
\item{signal.par}{a named list with elements `alpha` and `beta` supplying the
power-law parameters to simulate the common signal.}

\item{noise.par}{as \code{signal.par} the power-law parameters to simulate
the independent noise components.}

\item{nc}{the number of proxy records (`"cores"`) in the array.}

\item{nt}{the length of each record in the array.}
}
\value{
a list with `nc` elements where each is one simulated proxy record of
  length `nt`.
}
\description{
Simulate an array of proxy records with each record being composed of a
signal that is common to all the records in the array and an independent
noise component. Both components are simulated such that their spectra follow
a power-law of the form `alpha * f^(-beta)` where `f` denotes frequency.
}
\author{
Thomas Münch
}
