% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCI.R
\name{simSignalAndNoise}
\alias{simSignalAndNoise}
\title{Simulate a core array and estimate the signal and noise spectra}
\usage{
simSignalAndNoise(signal.par, noise.par, nc, nt, res, df.log = NULL)
}
\arguments{
\item{signal.par}{a named list with elements `alpha` and `beta` supplying the
power-law parameters to simulate the common signal.}

\item{noise.par}{as \code{signal.par} the power-law parameters to simulate
the independent noise components.}

\item{nc}{the number of proxy records (`"cores"`) in the array.}

\item{nt}{the length of each record in the array.}

\item{res}{the sampling (e.g., temporal) resolution of the simulated proxy
records; determines the frequency axis of the spectral estimates.}

\item{df.log}{width of the Gaussian kernel in logarithmic frequency units to
smooth the spectral estimates; \code{NULL} (the default) suppresses
smoothing.}
}
\value{
a named list of the three spectral objects `signal`, `noise` and
  `snr`, giving the spectra of the estimated common signal, the independent
  noise, and the corresponding signal-to-noise ratio (SNR), respectively.
}
\description{
This is a wrapper function to first simulate a proxy core array
(\code{\link{simCoreArray}}), estimate the corresponding array spectra
(\code{\link{ObtainArraySpectra}}), and finally separate for these simulated
data the common signal from the independent noise
(\code{\link{SeparateSignalFromNoise}}).
}
\author{
Thomas Münch
}
\keyword{internal}
