% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackCorrelation.R
\name{StackCorrelation}
\alias{StackCorrelation}
\title{Correlation with common signal}
\usage{
StackCorrelation(input, N, f1 = 2, f2 = "max", freq.cut.lower = NULL,
  freq.cut.upper = NULL)
}
\arguments{
\item{input}{a list of the spectral objects lists \code{signal} and
\code{noise}, usually to be obtained from a call to
\code{\link{SeparateSpectra}}}

\item{N}{maximum number of records in the assumed stack; correlations are
then calculated for stacks with record numbers from 1 to \code{N}}

\item{f1}{index of the the minimum frequency from which to integrate the
signal and noise spectra for calculating the correlation; per default the
lowest frequency of the spectral estimates is omitted}

\item{f2}{index of the maximum frequency until which to integrate the signal
and noise spectra for calculating the correlation; defaults to use the
maximum frequency of the given spectral estimates}

\item{freq.cut.lower}{lower frequency (not index!) at which to cut the
spectra: this provides a direct way for specifying a minimum frequency for
the integration different from the minimum frequency of the spectral
estimates. Setting \code{freq.cut.lower} overrides the frequency
corresponding to the index set in \code{f1}.}

\item{freq.cut.upper}{upper frequency (not index!) at which to cut the
spectra: this provides a direct way for specifying a maximum frequency for
the integration different from the maximum frequency of the spectral
estimates. Setting \code{freq.cut.upper} overrides the frequency
corresponding to the index set in \code{f2}.}
}
\value{
a list of three components:
\describe{
\item{signal:}{a spectral object list of the integrated signal}
\item{noise:}{a spectral object list of the integrated noise}
\item{correlation:}{a \code{n * m} matrix where \code{n} corresponds to
\code{N} and \code{m} is given by \code{length(f1 : f2)} providing the
correlation values as a function of the number of averaged records and
the record resolution (= increasing upper frequency of the integration)}
}
}
\description{
This function calculates the theoretical correlation of a \code{"stacked"}
proxy record with the common signal depending on the number of records in
the stack and the time resolution of the records, given estimates of the
average proxy signal and noise spectra.
}
\details{
The function is an implementation of Eqs. (6) and (7) in Münch and Laepple
(2018). The integral in (6) is approximated by the cumulative sum of the
integration arguments from \code{f.int1} to \code{f.int2}, where
\code{f.int1 = f1} and \code{f.int2} consecutively increases from \code{f1}
to \code{f2}.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
}
\author{
Thomas Münch
}
