% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenchNoise.R
\name{TrenchNoise}
\alias{TrenchNoise}
\title{Trench noise spectrum}
\usage{
TrenchNoise(trench.data, acc.rate = 25, sigma.acc.rate = 5, res = 3,
  neff = 19, df.log = 0.1)
}
\arguments{
\item{trench.data}{list of the shallow trench isotope profiles from Münch et
al. (2017)}

\item{acc.rate}{assumed mean accumulation rate at the trench site [cm/yr]}

\item{sigma.acc.rate}{assumed uncertainty of the accumulation rate [cm/yr]}

\item{res}{depth resolution of the trench data; defaults to 3 cm.}

\item{neff}{effective number of trench records to account for the spatial
autocorrelation of the stratigraphic noise; defaults to 19 for the total
available number of 22 profiles (Münch et al., 2017).}

\item{df.log}{Gaussian kernel width in log space to smooth the spectral
estimate; defaults to 0.1.}
}
\value{
A list of the three components:
\describe{
\item{\code{noise.lower}:}{spectral object list with the estimated trench
noise spectrum for the assumed lower bound of the accumulation rate.}
\item{\code{noise.mean}:}{as previous but for the assumed mean accumulation
rate.}
\item{\code{noise.upper}:}{as previous but for the assumed upper bound of
the accumulation rate.}
}
}
\description{
Internal function to obtain estimates of the noise spectrum at the local
scale at EDML using the available isotope data from the shallow trench
profiles from Münch et al. (2017). To convert the trench depth axis into a
time axis, a constant accumulation rate with a given uncertainty is
assumed.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
}
\author{
Thomas Münch
}
