% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wais}
\alias{wais}
\title{WAIS oxygen isotope data}
\format{
A data frame with 201 rows and 5 variables where each variable is the
time series of oxygen isotope data from one of the firn cores in per mil for
the years from 2000 CE to 1800 CE.
}
\source{
Data citation:\cr
Steig, E. J.: West Antarctica Ice Core and Climate Model Data, U.S. Antarctic
Program (USAP) Data Center, \url{https://doi.org/10.7265/N5QJ7F8B}, 2013.

License:\cr
CC-BY-NC-4.0 (\url{https://creativecommons.org/licenses/by-nc/4.0/})

Literature citation:\cr
Steig, E. J., et al.: Recent climate and ice-sheet changes in West Antarctica
compared with the past 2,000 years, Nat. Geosci., 6, 372–375,
\url{https://doi.org/10.1038/ngeo1778}, 2013.\cr
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, \url{https://doi.org/10.5194/cp-14-2053-2018},
2018.
}
\usage{
wais
}
\description{
A dataset containing the oxygen isotope data of 5 firn cores from the West
Antarctic Ice Sheet (WAIS), Antarctica, as used in Münch and Laepple (2018).
}
\details{
Original source data was downloaded from
\url{https://doi.org/10.7265/N5QJ7F8B} including the firn cores WDC2005A,
ITASE-1999-1, ITASE-2000-1, ITASE-2000-4 and ITASE-2000-5. The source data
was processed as documented in the package source
(\code{data-raw/load.WAIS.R}) which inludes the selection of the oxygen
isotope data and the required time span, bin-averaging to annual resolution
of the cores ITASE-2000-4 and ITASE-2000-5 from their original
higher-resolution data, and linear interpolation of missing values.
}
\keyword{datasets}
