% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-data.R
\docType{data}
\name{diffusion.tf}
\alias{diffusion.tf}
\title{Calculated diffusion transfer functions}
\format{A list with 3 elements:
\tabular{rll}{
  \code{dml1}: \tab a list with 2 numeric vectors of length 97: \tab  \cr
    \tab \code{freq}: frequency axis, in yr^{-1} \tab  \cr
    \tab \code{spec}: transfer function value at each frequency \tab  \cr
  \code{dml2}: \tab a list with 2 numeric vectors of length 497: \tab  \cr
    \tab \code{freq}: frequency axis, in yr^{-1} \tab  \cr
    \tab \code{spec}: transfer function value at each frequency \tab  \cr
  \code{wais}: \tab a list with 2 numeric vectors of length 100: \tab  \cr
    \tab \code{freq}: frequency axis, in yr^{-1} \tab  \cr
    \tab \code{spec}: transfer function value at each frequency \tab  \cr
}}
\source{
The transfer functions were obtained using \code{?DiffusionTF} for the
site-specific diffusion lengths provided by \code{?diffusion.length}. Isotope
records were simulated at semiannual resolution and the transfer functions
interpolated in frequency space to annual resolution. See also the respective
package vignette:
\code{vignette(topic = "diffusion-tf", package = "proxysnr")}.
}
\usage{
diffusion.tf
}
\description{
A dataset containing the calculated diffusion transfer functions for the
DML1, DML2 and WAIS data sets from Münch and Laepple (2018) based on
numerical simulations. The transfer function describes the average effect
diffusion has on the mean of the spectra from \code{N} spatially distributed
oxygen isotope records.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., \url{https://doi.org/10.5194/cp-2018-112}, in review,
2018.
}
\seealso{
\code{\link{DiffusionTF}}
}
\keyword{datasets}
