% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffusion.tf}
\alias{diffusion.tf}
\title{Calculated diffusion transfer functions}
\format{
A named list with the 3 elements \code{dml1}, \code{dml2}, and
  \code{wais}, where each element is a list of 2 numeric vectors:
  \describe{
  \item{\code{freq}:}{the frequency axis in units of \code{1 / yr},}
  \item{\code{spec}:}{the transfer function value at each frequency.}
}
}
\source{
The transfer functions were obtained using \code{?CalculateDiffusionTF} for
the site-specific diffusion lengths provided by
\code{proxysnr:::diffusion.length}. Isotope records were simulated at
semiannual resolution and the transfer functions interpolated in frequency
space to annual resolution. See also the respective package vignette:
\code{vignette(topic = "calculate-transfer-functions", package = "proxysnr")}.
}
\usage{
diffusion.tf
}
\description{
A dataset containing the calculated diffusion transfer functions for the
DML1, DML2 and WAIS data sets from Münch and Laepple (2018) based on
numerical simulations. The transfer function describes the average effect
diffusion has on the mean of the spectra from \code{N} spatially distributed
oxygen isotope records.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, \url{https://doi.org/10.5194/cp-14-2053-2018},
2018.
}
\seealso{
\code{\link{CalculateDiffusionTF}}\cr
\code{vignette(topic = "calculate-transfer-functions", package = "proxysnr")}
}
\keyword{datasets}
