% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{LLines}
\alias{LLines}
\title{Add spectrum to existing log-log spectral plot}
\usage{
LLines(x, conf = TRUE, bPeriod = FALSE, col = "black", alpha = 0.2,
  removeFirst = 0, removeLast = 0, ...)
}
\arguments{
\item{x}{an object of class \code{"spec"}.}

\item{conf}{if \code{TRUE} (the default) add a transparent confidence
interval (suppressed if \code{x} contains no error limits).}

\item{bPeriod}{if \code{TRUE} treat the x-axis values in units of period
(inverse frequency). Defaults to \code{FALSE}.}

\item{col}{color for the line plot and the confidence interval.}

\item{alpha}{transparency level (between 0 and 1) for the confidence
interval. Defaults to \code{0.2}.}

\item{removeFirst}{omit \code{removeFirst} values on the low frequency side.}

\item{removeLast}{omit \code{removeLast} values on the high frequency side.}

\item{...}{further graphical parameters passed to \code{lines}.}
}
\description{
This function adds a spectrum to an existing double-logarithmic plot and
optionally adds a transparent confidence interval.
}
\author{
Thomas Laepple
}
