% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-utils.R
\name{LPlot}
\alias{LPlot}
\title{Plot spectral object}
\usage{
LPlot(
  x,
  type = "l",
  inverse = FALSE,
  conf = TRUE,
  axes = TRUE,
  col = "black",
  alpha = 0.2,
  removeFirst = 0,
  removeLast = 0,
  xlab = ifelse(inverse, "period", "f"),
  ylab = "PSD",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{a spectral object.}

\item{type}{1-character string giving the type of plot desired: default type
"l" makes a line plot, while type "n" produces only the plot frame; see
also \code{\link{plot.default}}.}

\item{inverse}{if \code{TRUE} the x-axis is displayed in units of period
(inverse frequency), increasing to the left; defaults to \code{FALSE}.}

\item{conf}{if \code{TRUE} (the default) add a transparent confidence
interval (has no effect if \code{x} contains no confidence limits).}

\item{axes}{if \code{FALSE} the plotting of the x and y axes is
suppressed; defaults to \code{TRUE}.}

\item{col}{color for the line plot and the confidence interval.}

\item{alpha}{transparency level (between 0 and 1) for the confidence
interval; defaults to \code{0.2}.}

\item{removeFirst}{omit \code{removeFirst} values on the low frequency side.}

\item{removeLast}{omit \code{removeLast} values on the high frequency side.}

\item{xlab}{character string for labelling the x-axis.}

\item{ylab}{character string for labelling the y-axis.}

\item{xlim}{range of x-axis values; if \code{NULL} (the default) it is
calculated internally and automatically reversed if \code{period = TRUE}.}

\item{ylim}{range of y-axis values; if \code{NULL} (the default) it is
calculated internally.}

\item{...}{further graphical parameters passed to \code{plot}.}
}
\description{
Plot a spectral object on a double-logarithmic scale, optionally adding a
transparent confidence interval.
}
\seealso{
`?spec.object` for the definition of a "proxysnr" spectral object.
}
\author{
Thomas Laepple
}
