% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting.R
\name{muench_laepple_fig05}
\alias{muench_laepple_fig05}
\title{Figure 5 in Münch and Laepple (2018)}
\usage{
muench_laepple_fig05(SNR, TNS = t15.noise, f.cut = FALSE)
}
\arguments{
\item{SNR}{output from \code{\link{PublicationSNR}}.}

\item{TNS}{T15 trench noise spectra; per default loaded from internal data.}

\item{f.cut}{Shall the spectra be cut at the cutoff frequency constrained
by the diffusion correction strength? Defaults to \code{FALSE}.}
}
\description{
Plot comparison of the DML and T15 trench noise spectra (Figure 5 in Münch
and Laepple, 2018).
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{PublicationSNR}}
}
\author{
Thomas Münch
}
