% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObtainArraySpectra.R
\name{ObtainArraySpectra}
\alias{ObtainArraySpectra}
\title{Spectral estimates of core array}
\usage{
ObtainArraySpectra(cores, res = 1, neff = length(cores), df.log = NULL, ...)
}
\arguments{
\item{cores}{a list of the proxy data from the core array. Each component is
expected to be a numeric vector of the proxy values of a common length.}

\item{res}{the sampling (e.g., temporal) resolution of the proxy data;
determines the frequency axis of the spectral estimates.}

\item{neff}{the effective number of records (e.g. to account for an expected
spatial correlation of the local noise). Per default, no spatial correlation
is assumed and \code{neff} is set to the number of proxy records (the length
of \code{cores}).}

\item{df.log}{width of the Gaussian kernel in logarithmic frequency units to
smooth the spectral estimates; \code{NULL} (the default) suppresses
smoothing.}

\item{...}{additional parameters which are passed to the spectral estimation
function \code{\link{SpecMTM}}.}
}
\value{
A list of the following components:
\describe{
\item{N:}{the number of (effective) proxy records of the core array.}
\item{single:}{a list of length \code{N} with the spectra of each individual
proxy record.}
\item{mean:}{the mean spectrum across all individual spectra.}
\item{stack:}{the spectrum of the average proxy record in the time domain
("stacked record").}
}
}
\description{
Calculate all relevant spectral estimates for a given array of \code{n}
proxy records. The spectral estimates can be smoothed in logarithmic space.
}
\details{
The spectral estimates are calculated using Thomson’s multitaper method with
three windows with linear detrending before analysis
(see \code{\link{SpecMTM}}). Each spectral result is returned as an object
of class \code{"spec"} with the minimum elements \code{freq} and
\code{spec}.
}
\seealso{
\code{\link{SpecMTM}}
}
\author{
Thomas Münch
}
