% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{PlotArraySpectra}
\alias{PlotArraySpectra}
\title{Plot proxy array spectra (Figure 1 in Münch and Laepple, 2018)}
\usage{
PlotArraySpectra(
  spec,
  f.cutoff = NA,
  xlim = c(100, 2),
  ylim = c(0.005, 50),
  col = c("darkgrey", "black", "burlywood4"),
  col.sn = c("dodgerblue4", "firebrick4"),
  alpha.sn = 0.2,
  plt.ann = "default",
  xlab = NULL,
  ylab = NULL,
  xtm = NULL,
  ytm = NULL,
  xtl = NULL,
  ytl = NULL
)
}
\arguments{
\item{spec}{output from \code{\link{ArraySpectra}}.}

\item{f.cutoff}{optional frequency value to draw a vertical line symbolising
a cutoff frequency.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits (y1, y2) of the plot.}

\item{col}{a three-element vector with the colours for the individual
spectra (\code{col[1]}), the mean (\code{col[2]}) and the stack spectrum
(\code{col[3]}).}

\item{col.sn}{a two-element vector with the colours for the approximate
signal (\code{col[1]}) and noise shading (\code{col[2]}).}

\item{alpha.sn}{opacity factor for the colours in \code{col.sn} within
[0,1].}

\item{plt.ann}{if \code{"default"} use axis annotation as in Fig. 1 of Münch
and Laepple (2018). Since no other fixed annotation scheme is implemented,
setting \code{plt.ann} to a different value will result in an error.}

\item{xlab}{if not \code{NULL} use this specific x axis label to override
the default setting.}

\item{ylab}{if not \code{NULL} use this specific y axis label to override
the default setting.}

\item{xtm}{if not \code{NULL} use this vector of specific x axis tick mark
positions to override the default setting.}

\item{ytm}{if not \code{NULL} use this vector of specific y axis tick mark
positions to override the default setting.}

\item{xtl}{x axis tick mark labels. If \code{xtm} is not \code{NULL},
\code{xtl} is used for the \code{labels} parameter of the \code{\link{axis}}
function: if \code{xtl} is set to \code{NULL}, the labels are determined
automatically, else set it explicitly by specifying a vector of labels of
the same length as \code{xtm}; see also \code{?axis}.}

\item{ytl}{equivalent to \code{xtl} for the y axis tick mark labels.}
}
\description{
Plot the spectral estimates from a spatial proxy core array.
}
\examples{
# Plot Figure 1 in Münch and Laepple (2018) (DML1 oxygen isotope data set):

PlotArraySpectra(ArraySpectra(dml$dml1, df.log = 0.12))
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{ArraySpectra}}
}
\author{
Thomas Münch
}
