% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotArraySpectra}
\alias{PlotArraySpectra}
\title{Plot spectra from spatial proxy data array}
\usage{
PlotArraySpectra(
  spec,
  marker = NA,
  xlim = c(100, 2),
  ylim = c(0.005, 50),
  col = c("darkgrey", "black", "burlywood4"),
  col.sn = c("dodgerblue4", "firebrick4"),
  alpha.sn = 0.2,
  xlab = "Time period (yr)",
  ylab = "Power spectral density",
  xtm = NULL,
  ytm = NULL,
  xtl = NULL,
  ytl = NULL
)
}
\arguments{
\item{spec}{output from \code{\link{ObtainArraySpectra}}.}

\item{marker}{vector of optional frequency values to mark certain parts of
the plot, e.g. a cutoff frequency, in form of vertical grey dashed lines.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits (y1, y2) of the plot.}

\item{col}{a three-element vector with the colours for the individual
spectra (\code{col[1]}), the mean (\code{col[2]}) and the stack spectrum
(\code{col[3]}).}

\item{col.sn}{a two-element vector with the colours for the approximate
signal (\code{col[1]}) and noise shading (\code{col[2]}).}

\item{alpha.sn}{opacity factor for the colours in \code{col.sn} within
[0,1].}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{xtm}{x axis tick mark positions; if \code{NULL} computed by
\code{\link[graphics]{axis}}.}

\item{ytm}{y axis tick mark positions; if \code{NULL} computed from the
\code{ylim} range in steps of powers of 10.}

\item{xtl}{x axis tick mark labels; if \code{NULL} determined automatically
from \code{xtm}, else it must be a vector of labels of the same length as
\code{xtm}.}

\item{ytl}{equivalent to \code{xtl} for the y axis tick mark labels.}
}
\description{
Plot the spectral estimates from a spatial proxy data array (e.g., as in the
firn core analysis of Münch and Laepple, 2018, Fig. 1).
}
\examples{

# Plot Figure 1 in Münch and Laepple (2018) (DML1 oxygen isotope data set):
PlotArraySpectra(ObtainArraySpectra(dml$dml1, df.log = 0.12),
                 ylab = expression(
                   "Power spectral density " * "(\u2030"^{2}\%.\%"yr)"))

}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{ObtainArraySpectra}}
}
\author{
Thomas Münch
}
