% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotSNR}
\alias{PlotSNR}
\title{Plot proxy signal-to-noise ratios}
\usage{
PlotSNR(
  spec,
  f.cut = FALSE,
  names = NULL,
  col = 1:length(spec),
  xlim = c(500, 2),
  ylim = c(0.05, 5),
  plt.ann = "default",
  xlab = NULL,
  ylab = NULL,
  xtm = NULL,
  ytm = NULL,
  xtl = NULL,
  ytl = NULL
)
}
\arguments{
\item{spec}{a (named) list of signal-to-noise ratio data sets: each data set
itself should be list containing at least a named element \code{snr} which
is an object of class \code{"spec"} providing signal-to-noise ratios as a
function of frequency. For Figure 3 in Münch and Laepple (2018) set
\code{spec} to the output from \code{\link{PublicationSNR}}.}

\item{f.cut}{Shall the spectra be cut at the cutoff frequency constrained
by the diffusion correction strength? Defaults to \code{FALSE}.}

\item{names}{an optional character vector of names of the proxy data
sets. If \code{NULL}, the names of \code{spec} are used or, if not present,
default names.}

\item{col}{a numeric or character vector of colors to use for the plotting
with length recycled to match \code{length(spec)}.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits (y1, y2) of the plot.}

\item{plt.ann}{if \code{"default"} use axis annotation as in Fig. 3 of Münch
and Laepple (2018). Since no other fixed annotation scheme is implemented,
setting \code{plt.ann} to a different value will result in an error.}

\item{xlab}{if not \code{NULL} use this specific x axis label to override
the default setting.}

\item{ylab}{if not \code{NULL} use this specific y axis label to override
the default setting.}

\item{xtm}{if not \code{NULL} use this vector of specific x axis tick mark
positions to override the default setting.}

\item{ytm}{if not \code{NULL} use this vector of specific y axis tick mark
positions to override the default setting.}

\item{xtl}{x axis tick mark labels. If \code{xtm} is not \code{NULL},
\code{xtl} is used for the \code{labels} parameter of the \code{\link{axis}}
function: if \code{xtl} is set to \code{NULL}, the labels are determined
automatically, else set it explicitly by specifying a vector of labels of
the same length as \code{xtm}; see also \code{?axis}.}

\item{ytl}{equivalent to \code{xtl} for the y axis tick mark labels.}
}
\description{
Plot the timescale dependence of proxy signal-to-noise ratios, such as in
Fig. 3 in Münch and Laepple, 2018.
}
\examples{
# Plot Figure 3 in Münch and Laepple (2018)
# (DML and WAIS oxygen isotope data sets):

# Load main spectral results
DWS <- WrapSpectralResults(
       dml1 = dml$dml1, dml2 = dml$dml2, wais = wais,
       diffusion = diffusion.tf,
       time.uncertainty = time.uncertainty.tf,
       df.log = c(0.15, 0.15, 0.1))

# Calculate the final signal-to-noise ratio spectra
SNR <- proxysnr:::PublicationSNR(DWS$dml1$corr.full, DWS$dml2$corr.full,
                                 DWS$wais$corr.full)

# Plot it
PlotSNR(SNR, f.cut = TRUE,
        names = c("DML", "WAIS"),
        col = c("black", "dodgerblue4"))
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{PublicationSNR}}
}
\author{
Thomas Münch
}
