% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{PlotStackCorrelation}
\alias{PlotStackCorrelation}
\title{Plot proxy stack correlation (Figure 4 in Münch and Laepple, 2018)}
\usage{
PlotStackCorrelation(
  freq,
  correlation,
  col.pal,
  n = nrow(correlation),
  label = "",
  plt.ann = "default",
  xlab = NULL,
  ylab = NULL,
  xtm = NULL,
  ytm = NULL,
  xtl = NULL,
  ytl = NULL,
  xtm.min = NULL,
  ytm.min = NULL,
  xlim = NA,
  ylim = NA
)
}
\arguments{
\item{freq}{frequency axis of the underlying proxy data set to obtain an
axis for the temporal averaging period; its length must match
\code{ncol(correlation)}.}

\item{correlation}{a \code{n * m} matrix of correlation values where the
number of columns, \code{m}, corresponds to the number of frequency values
at which the correlation has been evaluated, i.e. \code{length(freq)}, and
the number of rows, \code{n}, to the total number of records averaged.}

\item{col.pal}{a color palette function to be used to assign colors in the
plot.}

\item{n}{the total number of records averaged; thus, the correlation map is
shown for records averaged from 1 to \code{n}. Per default set to the number
of rows in \code{correlation}.}

\item{label}{an optional label of the data set to be displayed at the top
of the plot.}

\item{plt.ann}{if \code{"default"} use axis annotation as in Fig. 4 of Münch
and Laepple (2018). Since no other fixed annotation scheme is implemented,
setting \code{plt.ann} to a different value will result in an error.}

\item{xlab}{if not \code{NULL} use this specific x axis label to override
the default setting.}

\item{ylab}{if not \code{NULL} use this specific y axis label to override
the default setting.}

\item{xtm}{if not \code{NULL} use this vector of specific x axis tick mark
positions to override the default setting.}

\item{ytm}{if not \code{NULL} use this vector of specific y axis tick mark
positions to override the default setting.}

\item{xtl}{x axis tick mark labels. If \code{xtm} is not \code{NULL},
\code{xtl} is used for the \code{labels} parameter of the \code{\link{axis}}
function: if \code{xtl} is set to \code{NULL}, the labels are determined
automatically, else set it explicitly by specifying a vector of labels of
the same length as \code{xtm}; see also \code{?axis}.}

\item{ytl}{equivalent to \code{xtl} for the y axis tick mark labels.}

\item{xtm.min}{if not \code{NULL} use these specific x axis minor tick marks
to override the default setting. Set to \code{NA} to omit minor ticks at
all.}

\item{ytm.min}{as \code{xtm.min} for minor y axis tick marks.}

\item{xlim}{the x limits (x1, x2) of the plot. Set to \code{NA} to use
default limits, or supply a numeric vector of length 2 with custom
limits in log units. In the latter case, setting either of the elements to
\code{NA} results in using the default limit for this element only.}

\item{ylim}{as \code{xlim} for the y limits of the plot.}
}
\description{
Plot the correlation of the spatial average of a certain number of proxy
records with the underlying common signal depending on the number of records
averaged and their temporal resolution.
}
\examples{
# Plot Figure 5 in Münch and Laepple (2018)
# (DML and WAIS oxygen isotope data sets):

# Load main spectral results
DWS <- WrapSpectralResults(
       dml1 = dml$dml1, dml2 = dml$dml2, wais = wais,
       diffusion = diffusion.tf,
       time.uncertainty = time.uncertainty.tf,
       df.log = c(0.15, 0.15, 0.1))

# Calculate the final signal-to-noise ratio spectra
SNR <- proxysnr:::PublicationSNR(DWS$dml1$corr.full, DWS$dml2$corr.full,
                                 DWS$wais$corr.full)

# Calculate the correlations
crl <- StackCorrelation(SNR$dml, N = 1 : 20,
       freq.cut.lower = 1 / 100,
       freq.cut.upper = SNR$dml$f.cutoff[2])

# Plot it
library(RColorBrewer)
palette <- colorRampPalette(rev(RColorBrewer::brewer.pal(10, "RdYlBu")))
PlotStackCorrelation(freq = crl$freq, correlation = crl$correlation,
          col.pal = palette, label = "DML", ylim = c(NA, log(50)))
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{StackCorrelation}}
}
\author{
Thomas Münch
}
