% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{PlotTF}
\alias{PlotTF}
\title{Plot transfer functions (Figure B1 in Münch and Laepple, 2018)}
\usage{
PlotTF(dtf = NULL, ttf = NULL, names = NULL, col = NULL,
  dtf.threshold = NULL, xlim = c(500, 2), ylim1 = c(0.005, 5),
  ylim2 = c(0.2, 1.5), plt.ann = "default", xlab = NULL, ylab1 = NULL,
  ylab2 = NULL, xtm = NULL, ytm1 = NULL, ytm2 = NULL, xtl = NULL,
  ytl1 = NULL, ytl2 = NULL)
}
\arguments{
\item{dtf}{A list of transfer function data sets: each data set is an object
of class \code{"spec"} (see \code{?spectrum}) with minimum components
\code{freq} and \code{spec}, or simply a named list with these two, where
component \code{freq} is a numeric vector providing a frequency axis and
component \code{spec} a numeric vector with the corresponding diffusion
transfer function values. If \code{NULL} (the default), the diffusion
transfer function provided with the package is plotted, which corresponds to
Figure B1 in Münch and Laepple (2018).}

\item{ttf}{As \code{dtf} but providing time uncertainty transfer
functions. If \code{NULL} (the default), the time uncertainty transfer
function provided with the package is plotted, which corresponds to Figure
B1 in Münch and Laepple (2018).}

\item{names}{an optional character vector of names for the transfer function
data sets. If \code{NULL}, the names of \code{dtf} and \code{ttf} are used
or, if not present, default names. If the diffusion and time uncertainty
data sets differ in number, you can provide a list of two vectors of names.}

\item{col}{a numeric or character vector of colors to use for the plotting;
if \code{NULL} default colors are used.}

\item{dtf.threshold}{optional critical diffusion transfer function
value to plot a corresponding horizontal line and vertical lines of
corresponding frequency cutoff values (omitted for \code{NULL}).}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim1}{the y limits (y1, y2) of the diffusion transfer function plot.}

\item{ylim2}{the y limits (y1, y2) of the time uncertainty transfer function
plot.}

\item{plt.ann}{if \code{"default"} use axis annotation as in Fig. B1 of
Münch and Laepple (2018). Since no other fixed annotation scheme is
implemented, setting \code{plt.ann} to a different value will result in an
error.}

\item{xlab}{if not \code{NULL} use this specific x axis label to override
the default setting.}

\item{ylab1}{if not ‘NULL’ use this specific y axis label for the first plot
to override the default setting.}

\item{ylab2}{if not ‘NULL’ use this specific y axis label for the second
plot to override the default setting.}

\item{xtm}{if not \code{NULL} use this vector of specific x axis tick mark
positions to override the default setting.}

\item{ytm1}{as \code{xtm} for the first y axis.}

\item{ytm2}{as \code{xtm} for the second y axis.}

\item{xtl}{x axis tick mark labels. If \code{xtm} is not \code{NULL},
\code{xtl} is used for the \code{labels} parameter of the \code{\link{axis}}
function: if \code{xtl} is set to \code{NULL}, the labels are determined
automatically, else set it explicitly by specifying a vector of labels of
the same length as \code{xtm}; see also \code{?axis}.}

\item{ytl1}{as \code{xtl} for the first y axis.}

\item{ytl2}{as \code{xtl} for the second y axis.}
}
\description{
Plot the spectral transfer functions of the effects of diffusion and time
uncertainty.
}
\examples{
# Plot Figure B1 in Münch and Laepple (2018), i.e. the used transfer
# functions to correct the DML and WAIS isotope spectra:

PlotTF(names = c("DML1", "DML2", "WAIS"), dtf.threshold = 0.5,
       col = c("black", "firebrick", "dodgerblue"))
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
}
\author{
Thomas Münch
}
