% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotTF}
\alias{PlotTF}
\title{Plot transfer functions}
\usage{
PlotTF(
  dtf = NULL,
  ttf = NULL,
  names = NULL,
  col = NULL,
  dtf.threshold = NULL,
  xlab = "Time period (yr)",
  ylab1 = expression(bar(G)),
  ylab2 = expression(Phi),
  xlim = c(500, 2),
  ylim1 = c(0.005, 5),
  ylim2 = c(0.2, 1.5),
  xtm = NULL,
  ytm1 = NULL,
  ytm2 = NULL,
  xtl = NULL,
  ytl1 = NULL,
  ytl2 = NULL
)
}
\arguments{
\item{dtf}{a list of spectral objects (`?spec.object`) of the transfer
function datasets that describe a diffusion-type smoothing process. See
Details for the meaning of \code{NULL}.}

\item{ttf}{as \code{dtf} but providing time uncertainty transfer
functions. See Details for the meaning of \code{NULL}.}

\item{names}{an optional character vector of names for the transfer function
datasets. If \code{NULL}, the names of \code{dtf} and \code{ttf} are used
or, if not present, default names. Provide a list of two vectors of names,
if the diffusional smoothing and time uncertainty datasets differ in
number, or, if they do **not** differ in number but belong to different
underlying datasets.}

\item{col}{a numeric or character vector of colors to use for the plotting;
if \code{NULL} default colors are used.}

\item{dtf.threshold}{optional critical diffusion transfer function
value to plot a corresponding horizontal line and vertical lines of
corresponding frequency cutoff values (omitted for \code{NULL}).}

\item{xlab}{x axis label.}

\item{ylab1}{y axis label for the diffusion transfer function.}

\item{ylab2}{y axis label for the time uncertainty transfer function.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim1}{the y limits (y1, y2) of the diffusion transfer function plot.}

\item{ylim2}{the y limits (y1, y2) of the time uncertainty transfer function
plot.}

\item{xtm}{x axis tick mark positions; if \code{NULL} computed by
\code{\link[graphics]{axis}}.}

\item{ytm1}{y axis tick mark positions on the diffusion transfer function
plot; default setting (\code{NULL}) uses
\code{c(0.01, 0.05, 0.1, 0.5, 1, 5)}.}

\item{ytm2}{y axis tick mark positions on the time uncertainty transfer
function plot; default setting (\code{NULL}) uses
\code{c(0.2, 0.4, 0.6, 0.8, 1, 1.2)}.}

\item{xtl}{x axis tick mark labels; if \code{NULL} determined automatically
from \code{xtm}, else it must be a vector of labels of the same length as
\code{xtm}.}

\item{ytl1}{y axis tick mark labels on the diffusion transfer function plot;
if \code{NULL} determined automatically from \code{ytm1}, else it must be a
vector of labels of the same length as \code{ytm1}.}

\item{ytl2}{as \code{ytl1} for the y axis on the time uncertainty transfer
function plot.}
}
\description{
Plot the spectral transfer functions of the effects of diffusion-like
smoothing and/or time uncertainty.
}
\details{
You can plot only transfer functions describing diffusion-like smoothing
processes by supplying the respective data to the \code{dtf} parameter and
leaving the \code{ttf} parameter as \code{NULL}, or vice versa for plotting
only time uncertainty transfer functions. Supplying both parameters will plot
the supplied diffusional smoothing as well as time uncertainty transfer
functions (which can differ in number) on a single plot with a shared x
axis. Leaving both the \code{dtf} and \code{ttf} parameters as \code{NULL}
(the default) will plot the firn diffusion and time uncertainty transfer
functions that are provided with the \code{proxysnr} package, which
corresponds to Figure B1 in Münch and Laepple (2018).
}
\examples{

# Plot Figure B1 in Münch and Laepple (2018), i.e. the used transfer
# functions to correct the DML and WAIS isotope spectra:

PlotTF(names = c("DML1", "DML2", "WAIS"), dtf.threshold = 0.5,
       col = c("black", "firebrick", "dodgerblue"))

}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
`?spec.object` for the definition of a "proxysnr" spectral object.
}
\author{
Thomas Münch
}
