% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublicationSNR.R
\name{PublicationSNR}
\alias{PublicationSNR}
\title{Final DML and WAIS spectra}
\usage{
PublicationSNR(spec, dml.knit.f = 0.1, df.log = 0.125)
}
\arguments{
\item{spec}{output from \code{\link{WrapSpectralResults}} for DML and WAIS
data}

\item{dml.knit.f}{frequency at which to combine the spectra from the DML1
and DML2 data sets (defaults to 1/10 yr^(-1)); DML2 spectra are used for
lower frequencies than \code{dml.knit.f}, DML1 for the higher frequencies.}

\item{df.log}{Gaussian kernel width in log space for additional smoothing
for visual purposes.}
}
\value{
A list with the components \code{dml} and \code{wais}, where each is
a list containing three elements:
\describe{
\item{\code{signal}:}{the signal spectrum.}
\item{\code{noise}:}{the noise spectrum.}
\item{\code{snr}:}{the frequency-dependent signal-to-noise ratio.}
}
}
\description{
Internal function to produce the final DML and WAIS spectra (signal, noise,
SNR) used for Figs. (3) and (4) in Münch and Laepple (2018). The DML spectra
are a combination of the results from the DML1 and DML2 data sets.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
}
\author{
Thomas Münch
}
