% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCI.R
\name{runSimulation}
\alias{runSimulation}
\title{Run simulation for the CI estimation}
\usage{
runSimulation(spectra, f.start = NULL, f.end = NULL, nmc = 10, df.log = NULL)
}
\arguments{
\item{spectra}{a list with the spectral objects \code{signal}, \code{noise},
and \code{snr} for an investigated proxy record array, obtained from
\code{\link{SeparateSignalFromNoise}}.}

\item{f.start}{lower end of the frequency range on which the power-law fit is
made on the proxy data (see Details); the default \code{NULL} uses the
lowest frequency of the proxy \code{spectra}.}

\item{f.end}{as \code{f.start} for the upper end; the default \code{NULL}
uses the uppermost frequency of the proxy \code{spectra}.}

\item{nmc}{integer; the number of replications for the confidence interval
estimation.}

\item{df.log}{width of the Gaussian kernel in logarithmic frequency units to
smooth the spectral estimates of the simulated data; some smoothing is
usually necessary to avoid physically implausible negative power
occasionally occuring for some frequencies upon estimating the common
signal spectrum. It is suggested to use the same amount of smoothing as for
the actual proxy data, while setting \code{NULL} (the default) suppresses
smoothing.}
}
\value{
the output of \code{runSurrogates}: a list of length \code{nmc}
  where each list element is one signal and noise estimation realization.
}
\description{
Get the required parameters from the input spectra (number of observations,
observation resolution, number of proxy records, and signal and noise
powerlaw parameters) and call the Monte Carlo signal and noise simulation
\code{\link{runSurrogates}}.
}
\seealso{
\code{\link{runSurrogates}}, \code{\link{EstimateCI}}
}
\author{
Thomas Münch
}
\keyword{internal}
