% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateSignalFromNoise.R
\name{SeparateSignalFromNoise}
\alias{SeparateSignalFromNoise}
\title{Calculate signal and noise spectra}
\usage{
SeparateSignalFromNoise(
  spectra,
  neff = spectra$N,
  diffusion = NULL,
  time.uncertainty = NULL
)
}
\arguments{
\item{spectra}{a list of the raw spectral estimates from a proxy core
array. Expected is the output from \code{\link{ObtainArraySpectra}}, but
sufficient is a named list of two components giving the \code{mean} and
\code{stack} spectra.}

\item{neff}{the effective number of records (e.g. to account for an expected
spatial correlation of the local noise). Per default set to element \code{N}
in \code{spectra}, otherwise supply it explicitly here.}

\item{diffusion}{numeric vector of diffusion correction values (inverse
transfer function); must be of the same length as the spectral estimates in
\code{spectra}. The default `NULL` is to apply no correction.}

\item{time.uncertainty}{numeric vector of time uncertainty correction
values (inverse transfer function); must be of the same length as the
spectral estimates in \code{spectra}. The default `NULL` is to apply no
correction.}
}
\value{
A list of three components, each of class \code{"spec"}:
\describe{
\item{\code{signal}:}{the raw or corrected signal spectrum.}
\item{\code{noise}:}{the raw or corrected noise spectrum.}
\item{\code{snr}:}{the signal-to-noise ratio as calculated from the previous
components.}
}
}
\description{
Calculate the raw signal and noise spectra, and the corresponding
signal-to-noise ratio, as estimated from a core array of \code{n} proxy
records, and correct these, where applicable, for the effects of time
uncertainty and water vapour diffusion (relevant for firn and ice cores).
}
\details{
This function is an implementation of Eq. (4) in Münch and Laepple
(2018). While the diffusion correction is relevant only for diffusing
proxies such as stable isotopes from firn and ice cores, this function can
be applied to a large set of proxy data since only one of the two
correction functions, or none, need to be supplied; thus, e.g., it can also
be used for non-diffusing proxy data where only time uncertainty is
relevant, or for estimating raw signal and noise spectra by supplying no
correction functions at all.
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal- to centennial-scale isotope variations from Antarctic ice cores?
Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
}
\seealso{
\code{\link{ObtainArraySpectra}}
}
\author{
Thomas Münch
}
