% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-plotting-functions.R
\name{SetPlotPar}
\alias{SetPlotPar}
\title{Default plotting parameters}
\usage{
SetPlotPar(mar = c(5, 5, 0.5, 0.5), las = 1, cex.main = 1.5,
  cex.lab = 1.5, cex.axis = 1.25, ...)
}
\value{
A list of plotting parameters to be used with \code{par()}.
\itemize{
  \item mar = c(5, 5, 0.5, 0.5)
  \item las = 1
  \item cex.main = 1.5
  \item cex.lab = 1.5
  \item cex.axis = 1.25
  \item ...
}
}
\description{
This function returns a list of default plotting parameters specified as its
function arguments, which can then be set via a call to
\code{par}. Additional parameters can be specified via \code{...}. This
wrapper function provides a convenient way to set new graphics parameters
and save their old values for later restoring at the same time; see the
example.
}
\examples{
# Default settings
op <- par(SetPlotPar())
plot(1 : 10, xlab = "X title", ylab = "Y title", type = "l")
par(op)

# Change default line width and increase outer margin
op <- par(SetPlotPar(lwd = 2, oma = c(2, 2, 2, 2)))
plot(1 : 10, xlab = "X title", ylab = "Y title", type = "l")
par(op)
}
\seealso{
\code{\link{par}}
}
\author{
Thomas Münch
}
