% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapSpectralResults.R
\name{WrapSpectralResults}
\alias{WrapSpectralResults}
\title{Main results for Münch and Laepple (2018)}
\usage{
WrapSpectralResults(..., diffusion, time.uncertainty, df.log = 0.05,
  crit.diffusion = 2)
}
\arguments{
\item{...}{a comma separated list of named proxy data sets to analyse.}

\item{diffusion}{a list of inverse transfer functions to correct for the
effect of diffusion. The length of the list has to match the number of
provided data sets, thus, one transfer function per data set is assumed.}

\item{time.uncertainty}{similar to \code{diffusion} a list of inverse
transfer functions to correct for the effect of time uncertainty.}

\item{df.log}{a vector of Gaussian kernel widths in log space to smooth the
spectral estimates from each data set. If dimensions do not fit, its length
is recycled to match the number of data sets.}

\item{crit.diffusion}{maximum diffusion correction value to obtain cutoff
frequencies until which results are analysed to avoid large uncertainties at
the high-frequency end of the spectra.}
}
\value{
A list of \code{N} lists, where \code{N} is the number of provided
data sets and where each of these lists contains four elements:
\describe{
\item{\code{f.cutoff}:}{a two-element vector with the index and value of the
cutoff frequency.}
\item{\code{raw}:}{the raw signal, noise and corresponding SNR spectra.}
\item{\code{corr.t.unc.only}:}{the signal, noise and corresponding SNR
spectra after correction for the effect of time uncertainty.}
\item{\code{corr.full}:}{the signal, noise and corresponding SRN spectra
after correction for both the effects of diffusion and time uncertainty.}
}
}
\description{
This internal wrapper function is used to combine all main spectral results
presented and discussed in Münch and Laepple (2018) but it can also be used
to combine the results for other data sets.
}
\examples{
# Get main results of Münch and Laepple (2018)

# Load the isotope data sets
dml  <- proxysnr:::load.DML()
wais <- proxysnr:::load.WAIS()
# Load the inverse transfer (correction) functions
dtf <- proxysnr:::load.DTF()
ttf <- proxysnr:::load.TTF()

# Main paper results
results <- proxysnr:::WrapSpectralResults(dml1 = dml$dml1,
               dml2 = dml$dml2, wais = wais,
               diffusion = dtf,
               time.uncertainty = ttf,
               df.log = c(0.15, 0.15, 0.1))
}
\references{
Münch, T. and Laepple, T.: What climate signal is contained in
decadal to centennial scale isotope variations from Antarctic ice cores?
Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
}
\author{
Thomas Münch
}
